+++ tests/test_core.py
@@ -167,6 +167,23 @@ def test_message_file_unicode():
         assert u'Räuber Hotzenplotz' in content, content
 
 
+def test_plugin_module(caplog):
+    """
+    Check if using a module with dotted name also works.
+    """
+
+    with caplog.at_level(logging.DEBUG):
+
+        # Bootstrap the core machinery without MQTT
+        core_bootstrap(configfile=configfile)
+
+        # Signal mocked MQTT message to the core machinery for processing
+        send_message(topic='test/plugin-1', payload='{"name": "temperature", "value": 42.42}')
+
+        # Proof that the message has been routed to the "log" plugin properly
+        assert 'Plugin invoked' in caplog.text, caplog.text
+
+
 def test_xform_func(caplog):
     """
     Submit a message to the "log" plugin and proof